spec_resnet

+++ resnet1d.py
@@ -22,7 +22,7 @@ class MyDataset(Dataset):
         self.label = label
 
     def __getitem__(self, index):
-        return (torch.tensor(self.data[index], dtype=torch.float), torch.tensor(self.label[index], dtype=torch.long))
+        return (torch.tensor(self.data[index], dtype=torch.float), torch.tensor(self.label[index], dtype=torch.float))
 
     def __len__(self):
         return len(self.data)
